fix: use BigDecimal instead of double, refactoring

+++ domain/src/main/java/com/a65apps/vbabikov/studyapp/calculator/CalculatorEffect.kt
@@ -1,5 +1,7 @@
 package com.a65apps.vbabikov.studyapp.calculator
 
+import java.math.BigDecimal
+
 sealed class CalculatorEffect {
 
     data class Input(
@@ -8,15 +10,15 @@ sealed class CalculatorEffect {
         val count: Int
     ) : CalculatorEffect()
 
-    data class Add(val operand1: Double) : CalculatorEffect()
+    data class Add(val operand1: BigDecimal) : CalculatorEffect()
 
-    data class Subtract(val operand1: Double) : CalculatorEffect()
+    data class Subtract(val operand1: BigDecimal) : CalculatorEffect()
 
-    data class Multiply(val operand1: Double) : CalculatorEffect()
+    data class Multiply(val operand1: BigDecimal) : CalculatorEffect()
 
-    data class Divide(val operand1: Double) : CalculatorEffect()
+    data class Divide(val operand1: BigDecimal) : CalculatorEffect()
 
-    data class Result(val operand2: Double) : CalculatorEffect()
+    data class Result(val operand2: BigDecimal) : CalculatorEffect()
 
     object Clear : CalculatorEffect()
